# backend/bin/main/static/openapi/pesaguard-developer-v1.yaml
openapi: 3.1.0
info:
  title: PesaGuard Developer Platform API
  version: 0.1.0
  description: Implemented secure developer-platform foundation. OAuth, billing, production environments, webhooks, and financial execution are out of scope.
servers:
  - {url: 'http://localhost:8080', description: Local development}
security:
  - {bearerAuth: []}
paths:
...
  /api/v1/organization:
    get: {responses: {'200': {$ref: '#/components/responses/Data'}, '401': {$ref: '#/components/responses/Error'}}}
    post: {requestBody: {$ref: '#/components/requestBodies/CreateOrganization'}, responses: {'201': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
    patch: {requestBody: {$ref: '#/components/requestBodies/UpdateOrganization'}, responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
    delete: {responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/verify:
    post: {requestBody: {$ref: '#/components/requestBodies/VerifyOrganization'}, responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/suspend:
    post: {responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/restore:
    post: {responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/disable:
    post: {responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/ownership/transfer:
    post: {requestBody: {$ref: '#/components/requestBodies/TransferOwnership'}, responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/security-settings:
    get: {responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}}}
    put: {requestBody: {$ref: '#/components/requestBodies/UpdateSecuritySettings'}, responses: {'200': {$ref: '#/components/responses/Data'}, '400': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}}
...
  /api/v1/organization/members/{membershipId}/history:
    get: {parameters: [{$ref: '#/components/parameters/MembershipId'}], responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/members/{membershipId}/status:
    patch: {parameters: [{$ref: '#/components/parameters/MembershipId'}], requestBody: {$ref: '#/components/requestBodies/UpdateMemberStatus'}, responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/members/{membershipId}/role:
    patch: {parameters: [{$ref: '#/components/parameters/MembershipId'}], requestBody: {$ref: '#/components/requestBodies/ChangeMemberRole'}, responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/invitations:
    post: {requestBody: {$ref: '#/components/requestBodies/CreateInvitation'}, responses: {'201': {$ref: '#/components/responses/InvitationCreated'}, '400': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
    get: {responses: {'200': {$ref: '#/components/responses/Data'}, '403': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/invitations/accept:
    post: {security: [], requestBody: {$ref: '#/components/requestBodies/AcceptInvitation'}, responses: {'201': {$ref: '#/components/responses/AuthCreated'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/v1/organization/invitations/{invitationId}:
    delete: {parameters: [{$ref: '#/components/parameters/InvitationId'}], responses: {'204': {description: Revoked}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque}
  parameters:
...
    MembershipId: {name: membershipId, in: path, required: true, schema: {type: string, format: uuid}}
    InvitationId: {name: invitationId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
...
    CreateOrganization: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateOrganizationRequest'}}}}
    UpdateOrganization: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateOrganizationRequest'}}}}
    VerifyOrganization: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/VerifyOrganizationRequest'}}}}
    TransferOwnership: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TransferOwnershipRequest'}}}}
    UpdateSecuritySettings: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateSecuritySettingsRequest'}}}}
    CreateInvitation: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateInvitationRequest'}}}}
    AcceptInvitation: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AcceptInvitationRequest'}}}}
    UpdateMemberStatus: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateMemberStatusRequest'}}}}
    ChangeMemberRole: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangeMemberRoleRequest'}}}}
...
  responses:
    AuthCreated: {description: Created; access token is returned once, content: {application/json: {schema: {$ref: '#/components/schemas/AuthEnvelope'}}}}
...
    Data: {description: Successful response envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
    Error: {description: Structured client error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
...
    InvitationCreated: {description: Invitation created; the invitation token is returned once and only stored as a hash, headers: {Cache-Control: {schema: {type: string, example: no-store}}}, content: {application/json: {schema: {$ref: '#/components/schemas/InvitationCreatedEnvelope'}}}}
...
  schemas:
...
    CreateOrganizationRequest: {type: object, required: [name, type, metadata], properties: {name: {type: string, minLength: 2, maxLength: 120}, type: {type: string, enum: [DEVELOPER, ENTERPRISE, PARTNER]}, metadata: {type: object, description: Free-form organization metadata; key, depth and size limits are enforced and credential-like keys are rejected}}}
    UpdateOrganizationRequest: {type: object, required: [name, type, metadata], properties: {name: {type: string, minLength: 2, maxLength: 120}, type: {type: string, enum: [DEVELOPER, ENTERPRISE, PARTNER]}, metadata: {type: object}}}
    VerifyOrganizationRequest: {type: object, required: [reference], properties: {reference: {type: string, maxLength: 120}}}
    TransferOwnershipRequest: {type: object, required: [membershipId], properties: {membershipId: {type: string, format: uuid}}}
    UpdateSecuritySettingsRequest: {type: object, required: [allowedAuthMethods, sessionTtlMinutes, idleTimeoutMinutes, maxSessions, credentialMinLength, credentialMaxLength, mfaRequired, ipAllowlist, securityEventTypes], properties: {allowedAuthMethods: {type: array, items: {type: string, enum: [PASSWORD]}}, sessionTtlMinutes: {type: integer, minimum: 5, maximum: 10080}, idleTimeoutMinutes: {type: integer, minimum: 5, maximum: 10080}, maxSessions: {type: integer, minimum: 1, maximum: 100}, credentialMinLength: {type: integer, minimum: 12, maximum: 64}, credentialMaxLength: {type: integer, minimum: 24, maximum: 72}, mfaRequired: {type: boolean, description: Rejected while no MFA provider is configured}, ipAllowlist: {type: array, items: {type: string, description: IPv4/IPv6 address or CIDR range}}, securityEventTypes: {type: array, items: {type: string, enum: [LOGIN_FAILURE, MEMBERSHIP_CHANGED, SECURITY_SETTING_CHANGED, ORGANIZATION_LIFECYCLE, INVITATION_CREATED, INVITATION_ACCEPTED, INVITATION_REVOKED]}}}}
    CreateInvitationRequest: {type: object, required: [email, role], properties: {email: {type: string, format: email, maxLength: 320}, role: {type: string, enum: [ADMIN, DEVELOPER, VIEWER]}}}
    AcceptInvitationRequest: {type: object, required: [token, email, password], properties: {token: {type: string, maxLength: 512}, email: {type: string, format: email, maxLength: 320}, password: {type: string, format: password, minLength: 12, maxLength: 128}, displayName: {type: string, minLength: 2, maxLength: 120, description: Required only when the invited email does not yet have an account}}}
    UpdateMemberStatusRequest: {type: object, required: [status], properties: {status: {type: string, enum: [ACTIVE, SUSPENDED, REVOKED]}, reason: {type: string, maxLength: 500}}}
    ChangeMemberRoleRequest: {type: object, required: [role], properties: {role: {type: string, enum: [ADMIN, DEVELOPER, VIEWER]}}}
    InvitationCreatedEnvelope: {type: object, required: [data, requestId], properties: {data: {type: object, required: [id, email, role, token, expiresAt], properties: {id: {type: string, format: uuid}, email: {type: string, format: email}, role: {type: string}, token: {type: string, description: Returned only when the invitation is created; stored server-side as a hash}, expiresAt: {type: string, format: date-time}}}, requestId: {type: string, format: uuid}}}
...
    Envelope: {type: object, required: [data, requestId], properties: {data: {type: object}, requestId: {type: string, format: uuid}}}
    AuthEnvelope: {type: object, required: [data, requestId], properties: {data: {type: object, required: [accessToken, tokenType, expiresAt, user, organization], properties: {accessToken: {type: string}, tokenType: {type: string, example: Bearer}, expiresAt: {type: string, format: date-time}, user: {type: object}, organization: {type: object}}}, requestId: {type: string, format: uuid}}}
...
    ErrorEnvelope: {type: object, required: [error], properties: {error: {type: object, required: [code, message, requestId, timestamp, violations], properties: {code: {type: string}, message: {type: string}, requestId: {type: string, format: uuid}, timestamp: {type: string, format: date-time}, violations: {type: array, items: {type: object}}}}}}